refactor(quil-node): extract peer_info_publisher phase
- master_node/peer_info_publisher.rs: Phase 5b (PeerInfo broadcaster
  + KeyRegistry publisher with Ed448 signing) — spawn(args), takes a
  PeerInfoPublisherArgs struct. ~237 lines extracted. The OnceLock
  `pi_worker_manager` stays in run_master_node's outer scope and is
  passed in as worker_manager_cell.

main.rs: 5,069 → 4,862 lines.
