+++ crates/quil-node/src/master_node/mod.rs
@@ -2,4 +2,5 @@ pub(crate) mod engines;
 pub(crate) mod frame_pipeline;
 pub(crate) mod keys;
 pub(crate) mod networking;
+pub(crate) mod peer_info_publisher;
 pub(crate) mod storage;
